MDEV-37788 fix uintNkorr/intNstore macros to avoid unaligned access
Replaced macros with optimized inline functions, that use memcpy
instead of unaligned access (translates to single instruction
on architectures we support). Used compiler byteswap intrinsics
where appropriate.

Removed byte_order_generic.h as well as Intel specific headers
byte_order_generic_{x86|x86_64}.h

Added some casts to places where intNstore writes into "const char*".

+++ sql/sql_statistics.h
@@ -262,7 +262,7 @@ public:
       ((uint8 *) values)[i]= (uint8) (val * prec_factor());
       return;
     case DOUBLE_PREC_HB:
-      int2store(values + i * 2, val * prec_factor());
+      int2store(values + i * 2, (uint16)(val * prec_factor()));
       return;
     }
   }
